[wicd] Move EnsureServiceState() to Manager
This commit moves the EnsureServiceState function from winsvc.go to
manager.go, and makes it a method of the Manager interface. This is
being done as this function needs to be able to open dependencies
of the service being updated, this requires a handle to the service
manager. The usage of the manager handle will be done in a future
commit.
